Updated all SQL Statements, Unit test and Integration Tests

+++ src/main/java/com/napier/sem/City.java
@@ -10,6 +10,6 @@ public class City {
 
     public String District;
 
-    public int Population;
+    public long Population;
 
 }
